+++ Sources/Sentry/Public/SentryEvent.h
@@ -22,6 +22,13 @@ NS_SWIFT_NAME(Event)
 @property (nonatomic, strong) SentryMessage *message;
 
 /**
+ * The error of the event. This property adds convenience to access the error directly in
+ * beforeSend. This property is not serialized. Instead when preparing the event the SentryClient
+ * puts the error into exceptions.
+ */
+@property (nonatomic, strong) NSError *_Nullable error;
+
+/**
  * NSDate of when the event occured
  */
 @property (nonatomic, strong) NSDate *timestamp;
@@ -163,6 +170,15 @@ NS_SWIFT_NAME(Event)
  * @return SentryEvent
  */
 - (instancetype)initWithLevel:(enum SentryLevel)level NS_DESIGNATED_INITIALIZER;
+
+/**
+ * Initializes a SentryEvent with an NSError and sets the level to SentryLevelError.
+ *
+ * @param error The error of the event.
+ *
+ * @return The initialized SentryEvent.
+ */
+- (instancetype)initWithError:(NSError *)error;
 
 @end